:recycle: Move combinations to common
This can be used again in future and can be unit tested outside of the
day 7 context.
