+++ AOC/Days2024/Day07/Day07.cs
@@ -1,3 +1,5 @@
+using AOC.Common;
+
 namespace AOC.Days2024.Day07;
 
 public class Day07 : Day<List<(long, List<long>)>, long, List<(long, List<long>)>, long>
@@ -29,7 +31,7 @@ public class Day07 : Day<List<(long, List<long>)>, long, List<(long, List<long>)
 
     private static bool CanBeMadeTrue((long, List<long>) line, string[] characters)
     {
-        var combinations = GetCombinations(line.Item2.Count - 1, characters);
+        var combinations = Combinations.GetCombinations(line.Item2.Count - 1, characters);
         foreach (var combination in combinations)
         {
             var total = line.Item2[0];
@@ -54,50 +56,4 @@ public class Day07 : Day<List<(long, List<long>)>, long, List<(long, List<long>)
             "||" => long.Parse($"{value1}{value2}"),
             _ => throw new NotImplementedException(),
         };
-
-    public static List<string[]> GetCombinations(int length, string[] characters)
-    {
-        var combinations = new List<string[]>();
-
-        var possibleCombinations = Math.Pow(characters.Length, length);
-        for (var i = 0; i < possibleCombinations; i++)
-        {
-            var inBase = DecimalToBase(i, characters.Length).PadLeft(length, '0');
-            combinations.Add(inBase.Select(bit => characters[int.Parse(bit.ToString())]).ToArray());
-        }
-
-        return combinations;
-    }
-
-    // https://stackoverflow.com/questions/923771/quickest-way-to-convert-a-base-10-number-to-any-base-in-net
-    private static string DecimalToBase(long decimalNumber, int baseNumber)
-    {
-        const int bitsInLong = 64;
-        const string digits = "0123456789ABCDEFGHIJKLMNOPQRSTUVWXYZ";
-
-        if (baseNumber < 2 || baseNumber > digits.Length)
-            throw new NotImplementedException("Base must be >= 2 and <= " + digits.Length);
-
-        if (decimalNumber == 0)
-            return "0";
-
-        var index = bitsInLong - 1;
-        var currentNumber = Math.Abs(decimalNumber);
-        var charArray = new char[bitsInLong];
-
-        while (currentNumber != 0)
-        {
-            var remainder = (int)(currentNumber % baseNumber);
-            charArray[index--] = digits[remainder];
-            currentNumber /= baseNumber;
-        }
-
-        var result = new string(charArray, index + 1, bitsInLong - index - 1);
-        if (decimalNumber < 0)
-        {
-            result = "-" + result;
-        }
-
-        return result;
-    }
 }
